Change the Parse/Sema interaction for Objective-C collective literal elements, so that it
  (1) Performs the lvalue-to-rvalue conversion on each element, which will make property accesses work correctly, and
  (2) Actually return an invalid expression when something goes wrong, so that we don't end up building an invalid AST

When actually building a collection literal, convert each of the collection elements to the type actually expected by the method that will create the collection. CodeGen tends to ignore some of these issues, but it's far better to ensure that the AST is consistent and correct.

While doing this, simplify the collection literal construction logic by using early returns and documenting what we're actually doing.