Adjust the relative path from clang to llvm-gcc for Xcode repackaging.
This is a blatant hack to fix <rdar://problem/10660445>, where clang fails
to find llvm-gcc when compiling a 32-bit kext.  As of Onyx, clang is installed
to Xcode's default toolchain directory, while llvm-gcc has remained in
$(DEVELOPER_DIR)/usr.  The new relative path in this commit works for that
scenario.  It won't work if clang is installed in some other OS-specific
toolchain directory, but we're not going to be supporting 32-bit kexts in
any new OS builds, so that's not fatal (and this patch doesn't make the
situation any worse for the OS-specific toolchain case).