Introduce automatic refactoring to modern objc literal & subscripting syntax.
The tool is covered by rdar://10205772.

Use the refactoring code to also provide warnings + fixits when the source makes use
of the, now legacy, NSFoundation APIs (rdar://10688837). The warnings are
  -Wobjc-legacy-literal-construction
  -Wobjc-legacy-container-subscript

Also provide warnings + fixits for redundant calls to
  stringWithString:
  arrayWithArray:
  dictionaryWithDictionary:
with string/array/dictionary literals as parameters (rdar://10401960). The fixit removes
the call and leaves only the literal. The warning is
  -Wobjc-redundant-api-use

All the above warnings are under the -Wobjc-cocoa-api diagnostic group.