ARM64 ADD/SUB high half intrinsics prototype cleanup.
Per the revised (v2.1) ARM AdvSIMD intrinsics documentation, the first
operand of the vaddhn_high_* (and similar) intrinsics is half the number
of elements of the destination type.

This allows for natural chaining of the low and high half narrowing add
operations to build up the full register, without any typecasting, which
is a good thing. For example,
int8x16_t subhn2_16b(int16x8_t a, int16x8_t b) {
  int8x8_t c = vsubhn_s16(a, b);
  return vsubhn_high_s16(c, a, b);
}

rdar://10861725