Do not use "byval" for arguments that are passed by reference.  rdar://11421458
We still need to audit the arm64 ABI for recent changes, but at least with the
older version that we've been using so far, large aggregates are passed
indirectly by reference.  The "getIndirect" call used in those cases was
inadvertently picking up the default "byval" argument value of true.  Since
the arm64 backend doesn't support byval, this led to catastrophic results.