ARM64: vqsubb_[su]8 and vqsubh_[su]16 intrinsics codegen.
Since i8 and i16 aren't legal types, we codegen these are vector operations
with the scalar values as lane zero of an otherwise undef vector. The backend
will normally code-gen that like any other vector sqsub instruction, but fancy
isel patterns and/or a peephole should be able to map the operation to the
scalar instruction these intrinsics are intended to represent. The resultant
code quality improvement is likely very minor, however (mostly removing a copy
here and there), so it's questionable if it's worth it.

rdar://12026333