And end to end test for querying whether or not all or any lanes are zero or non-zero.
This is a C source test using ARM64 intrinsics.

Given the nature of the "if (max(…))" and "if (!max(…))" style of code, I
have made sure that the branch targets are correct as well, by capturing
the label being branched to, then CHECKing what follows that label.

<rdar://problem/11487757> ARM64: Smart isel for special case values in vector
compare and branch