Two tests for vector compare and branch.:
First, an update to relax the checks for an end to end test for querying
whether or not all or any lanes are zero or non-zero.  This is a C source
test using ARM64 intrinsics.

Second, a ".ll" version of the test, so people who only do "make check" will
find when something breaks.

In both cases, given the nature of the "if (max(…))" and "if (!max(…))"
style of code, I have made sure that the branch targets are correct as well,
by capturing the label being branched to, then CHECKing what follows that
label.

(Of course these two tests will be in separate check-ins, as they are in the
clang and llvm repos, respectively.)

<rdar://problem/11487757> ARM64: Smart isel for special case values in vector
compare and branch