Update the change for <rdar://problem/12617764> to only change the load size.
We decided to split this change into 2 steps to ease the integration.  Since
I had already applied the previous patch from Fariborz to the Osceola branch,
this change updates it to match what we ended up doing for the first step on
clang-500.2-branch.  Trunk has changed since we branched, so this isn't
completely the same as the version on the branch.