Change the toolchain_header_search test to use -cc1.
The -resource-dir option is only recently accepted by the clang driver
(and not in clang-500) and it doesn't really work. The driver seems to be
passing two -resource-dir options to the cc1 invocation. The first one has
the directory located relative to the clang binary, and the second one has no
argument at all.  The second one wins and the following argument ("-x") is
interpreted as the resource directory. The include path then just ignores the
non-existent "-x/include" directory, and somehow the test passes anyway.
Clearly there are some things that need further investigation here, but this
change will at least get the buildbots passing again.