[Frontend] Change implicit sysroot search behavior to always search SDK first.
 - <rdar://problem/14281015> unable to import Foundation from 10.9 SDK when
   adding explicit -F/S/L/F/ApplicationServices.framework/Frameworks

In light of how framework lookup works, it makes sense to always look in the SDK
first because if anything is found in the SDK it always makes sense for it to
take precedence. The only case where this would not be correct is if the user
was intentionally trying to override SDK content, but attempting to do so by
relying on header search order (or by explicitly overwriting content in the OS)
is the wrong way to do that and not something we need to support.