ARM64: move to OSS implementation of "no Q" type modifier. I.e. R -> H.
Hao recently added essentially the same concept as we have to the NEON
intrinsic generator, under the letter 'H' rather than 'R'. This means "the
type's 128-bit, but don't add a 'q' to the name, and is used in _high_
intrinsics (amongst others, it turns out).

Changing ARM64NeonEmitter.cpp to match reduces the divergences between it and
NeonEmitter.cpp substantially, and I think I actually prefer Hao's
implementation (the "qsuffix" classification is used in exactly one place, but
gets threaded everywhere).

Now if only I could undo the OSS decision to share AArch64 and AArch32
definitions. That was a (well, my) mistake (and is a mess). Unfortunately doing
so now would cause even more chaos on Osceola so I think it will have to wait
until we push it.