Move implicit-function-as-error test to the Driver directory.
The code being tested here, to automatically promote the warning to an error,
is in the driver, so it makes sense to test it along with other driver features.
But, the real reason to move it is that as of clang svn r199348, all the tests
in the Sema directory have to use clang_cc1.