Resolve merge correctly.
The problem was that there was a #ifndef __OPEN_SOURCE__ section in a piece of
code which had been removed in the open source code base, but was still left in
our internal sources.

I am moving the __OPEN_SOURCE__ section to the place in LLVM where this type of
information is being stored now.

Look for a complementary revision in llvm in the near future.