Add the APINotes library.
API notes are a mechanism for specifying additional information about declarations without having to modify the headers directly. For example, noting that a particular function always returns a non-null pointer. Semantic analysis will (eventually) query API notes to attach additional information (typically in the form of attributes) to declarations as they are type-checked.

API notes have two forms, a YAML source form that is easy to generate and manipulate and a bitcode-based binary format that’s optimized for efficient lookup. Add a driver tool -cc1apinotes that provides the YAML-to-binary and binary-to-YAML translation.