Use APINotes to add attributes to declarations during semantic analysis.
When -fapinotes is provided, automatically load API notes files that correspond to headers. API notes can be found in various places:
  - For a non-framework header, we search for the file APINotes.apinotes in the directory of the header and in its parent directories
  - For a public framework header (e.g., in the Headers directory), we search for the file APINotes/FrameworkName.apinotes within the .framework directory.
  - For a private framework header (e.g., in the PrivateHeaders directory), we search for the file APINotes/FrameworkName_private.apinotes in the .framework directory

Whenever we parse a declaration that can be annotated, we check whether there are any API notes that appertain to that declaration. We then map those API notes to attributes. In this initial implementation, only available information is mapped (to the “unavailable” attribute). More to come.

The API notes manager is compiling the YAML sources into the binary form, which is then used for the API notes search. The API notes manager should be caching the binary form (e.g., similarly to the way modules are cached, although the mechanism can be far simpler).