Disable -Werror when GCC_TREAT_WARNINGS_AS_ERRORS=NO is set.
XBS is going to append that setting to the build command for projects
that do not have approval to build with -Werror. For Xcode projects, that
will cause xcodebuild to disable -Werror, but for Makefile projects, we
need the compiler to handle it directly. Make will export that setting
from the command line into the environment, so that clang can check for
it. rdar://problem/17878132