ARMv7k: move back to unique RTTI
The implementation of non-unique RTTI relies on there being an unused high bit
in 64-bit pointers (bit 63, actually). This isn't true in general for 32-bit
systems, so it will need reworking if we want to support that on ARMv7k.

Slightly complicated by the fact that we *do* want to keep the revised tail
padding rules from ARM64, so we need a new iOSxyz ABI type.

rdar://problem/18890711