Re-revert most of r165273.
This change was reverted from our master branch due to rdar://problem/15055872.
The revert was accidentally lost when resolving a merge conflict: see
7683322a51a6b61946603c7234bc31fb143b4160.

However, the change cannot be completely reverted now without breaking
the UBSan tests. This change uses Kuba's approach for clang-602 where he
kept the check for the sanitizer but not the unreachable code that causes
the problematic optimizations. See rdar://problem/18890138.