Check for consistent use of nullability type specifiers in a header.
If any nullability type specifiers are used in a header, then warn about any pointer, block pointer, or member pointer declarators within that header that are not annotated with any nullability. Note that this warning will underrepresent the number of non-annotated pointer types in the case where more than one pointer declarator is seen before the first nullability type specifier, so that we need to only track a small amount of information (for efficient reasons). Part of rdar://problem/18868820.