Map context-sensitive nullability keywords to attributes within the parser.
Rather that applying the context-sensitive nullability keywords (in
Objective-C methods and properties) on top of the
return/parameter/property type during semantic analysis, map them to
(context-sensitive) attributes on the declarator or innermost
declarator chunk. This eliminates some redundant logic in Sema, treats
the context-sensitive nullability as the first type nullability
specifier for redundant/conflicting diagnostics, and eliminates false
positives in the nullability completeness warning.