Allow an Objective-C class to inherit from a specialized class type.
Permits, e.g., NSMutableArray<T> to inherit from NSArray<T>. As part
of this, generalize the representation of superclasses in Objective-C
interfaces (to a TypeSourceInfo*), but maintain the older entry points
to limit the refactoring at the moment.

Parsing the type arguments of the superclass is syntactically
ambiguous with the list of protocols to which the class conforms, so
we perform the same disambiguation that we use when parsing other
specialized types. Part of rdar://problem/6294649.