Substitute type arguments into the result type of Objective-C message sends.
When messaging a method that was defined in an Objective-C class (or
category or extension thereof) that has type parameters, substitute
the type arguments for those type parameters. This includes general
infrastructure for substituting the type arguments associated with an
ObjCObject(Pointer)Type into a type referenced within a particular
context, handling all of the substitutions required to deal with
(e.g.) inheritance involving parameterized classes. In cases where no
type arguments are available (e.g., because we're messaging via some
unspecialized type, id, etc.), we substitute in the type bounds for
the type parameters instead.

Example:

  @interface NSSet<T : id<NSCopying>> : NSObject <NSCopying>
  - (T)firstObject;
  @end

  void f(NSSet<NSString *> *stringSet, NSSet *anySet) {
    [stringSet firstObject]; // produces NSString*
    [anySet firstObject]; // produces id<NSCopying> (the bound)
  }