Refactor Objective-C type argument substitution into more useful APIs (NFC).
It was awkward for callers to have to dig down to an ObjCObjectType or
ObjCObjectPointerType to perform substitutions, particularly when one
can also deal with blocks (in some cases). Instead, provide the
substitution operation on QualType directly, taking the object type as
a parameter, so we encapsulate that knowledge. Additionally, provide
interfaces that allow one to separately compute and substitute type
arguments, so that future callers can ammortize the cost of computing
type arguments.