Remove Objective-C type arguments and protocol qualifiers from declaration specifiers.
Neither Objective-C type arguments nor protocol qualifiers are
actually declaration specifiers, yet they were modeled as part of
DeclSpec. Remove them from DeclSpec and instead provide specific Sema
entry points to form a specialized and/or protocol-qualified
type. This simplifies the handling of these qualifiers, because we no
longer need to distinguish between those type arguments and protocol
qualifiers that were written on the DeclSpec vs. being embedded in the
underlying type. It also properly diagnoses bogus types such as
"<NSCopying> id", which was treated as equivalent to "id<NSCopying>"
but should have been rejected (and now is) as a redundant type
specifier (the anchronistic spelling of "id<NSCopying>" followed by
"id").

One downside to this approach, which is also a downside to our
handling of templates, is that we end up copying the TypeLoc
information into the declarator, which can waste some memory in the
ASTContext.

Cleanup motivated by rdar://problem/6294649.