Add new -fembed-bitcode option.
This causes a copy of the optimized LLVM IR to be embedded in the generated
code in a separate data section. As a temporary workaround for not having all
the backend options represented as function attributes, it also embeds a copy
of the command-line options. It currently save all the options, but we're going
to pare that down using a whitelist of backend options that we know need to be
saved.

Based on a patch from Steven Wu. rdar://problem/18909437