Treat Objective-C __kindof types like id/Class in message sends.
When messaging a __kindof type, we consult the global method pool the
same way we would for message sends to id/Class, because we
dynamically treat a __kindof type as any potential subclass. This
implementation is a stepping stone toward rdar://problem/6985493; in
the future, we can use the class bound on the __kindof type to ignore
any methods declared within classes that aren't part of the hierarchy
of the class bound (neither superclass of, subclass of, nor the same
as the class bound).