Objective-C generics: substitute __kindof type bounds into unspecialized types.
When computing the type of a message send or property read for an
unspecialized parameterized class, substitute the __kindof form of
each type parameter's bound for the corresponding type parameter. This
preserves more type information that the previous solution, which
dropped all the way down to id/Class, losing too much type
information.

Part of the interaction between Objective-C generics and __kindof,
rdar://problem/6985493.