Allow MACOSX_DEPLOYMENT_TARGET to be set for TV OS and WatchOS.
For historical reasons, clang has always allowed both the
MACOSX_DEPLOYMENT_TARGET and IPHONEOS_DEPLOYMENT_TARGET env vars to be
set at the same time. It chooses between them based on whether the
architecture is ARM vs. Intel. For the new TV and Watch platforms, we had
hoped to avoid this hack, but those new platforms inherited their builds
from iOS and it's just not worth the trouble of trying to change it.
rdar://problem/19984882