Don't allow nullability qualifiers to move from the decl-specifiers to a pointer/block declarator.
Feedback from the open-source community has determined that this part
of the nullability semantics is problematic, so start to phase it out
(rdar://problem/20084364). This particular commit introduces a warning
(+ Fix-It) that fires when we need to move the nullabiltiy
qualifier. It's phrased as a warning that defaults to an error so we
can downgrade it where needed to keep things building while we make
the transition.