Add the swift_name attribute.
This attribute allows developers to control how their API is imported into
Swift, for the three major declaration kinds where the Swift name already
differs from the Objective-C name in a non-trivial way:

- Enum constants (prefix stripping)
- Protocols (may conflict with classes)
- Factory methods (may or may not turn into convenience initializers)

We enforce a particular schema on the argument string for each of these.
The former two must be simple identifiers; the latter must be a Swift-style
method name with the same number of parameters as the method.

rdar://problem/19240840