AArch64: remove accommodation for broken vfma/vfms semantics
Early versions of Clang had obviously broken semantics for vfma, swapping the
argument order compared to other intrinsics and the instructions for no valid
reason. Unfortunately, some code existed that relied on this broken behaviour
so we put a preprocessor flag in to help the transition. We can now remove it.
With extreme prejudice.

rdar://18075117