Workaround rdar://20754602 by bailing out early from highlightRange
We seem to have a backwards source range coming from "nonnull Foo *"
where we go from the Foo to the nonnull. The code that produces this
range looks wrong, but it would be risky to try a general fix.  So jsut
bail out early, since it's only diagnostic QOI issue if we don't
highlight the ranges.