Allow swift_name on classes, mainly for Swift's generated headers.
There are problems putting swift_name on arbitrary classes and protocols
because it interferes with lookup, but for interfaces generated by
Swift itself it only needs to match up the declaration with the
corresponding Swift declaration.

Because of the aforementioned problems, make the "swift_name cannot go
on this decl" error more vague, so that it doesn't promise features we
haven't implemented.

Part of rdar://problem/17469485