Allow more stack usage for test.
The test currently fails on the sanitizer bots because its stack usage
is huge. The overhead of the asan instrumentation generated by clang-700
on this test is much higher than what ToT generates (8x vs 3x).

Make the test pass, but we certainly need to understand why our overhead
is so high.