Bitcode: Preserve use-list order in the -cc1 serialization
For the -fembed-bitcode flow, we write the bitcode out twice:
  - in the first -cc1 call, and
  - to create the embedded bitcode in the second -cc1 call.
The second one should be removed going forward (rdar://21489435), but
for now just record the use-lists there so that we get the same
use-lists in both versions of the Archive build.

There isn't a good way to check in a test for this.

rdar://problem/20651301