Properly copy ObjCObjectTypeLocs that contain padding.
Fixes a rare crasher where we would copy four extra bytes of TypeLoc
information into ASTContext-allocated memory. The crasher was rare
because memory allocated to the destination TypeLoc would have to be
at the end of a page boundary (due to the underlying
BumpPtrAllocator). Note that the problem also (occasionally) manifests
as missing or incorrect source-location information due to padding
bytes being copied, hence the token-annotation test. Fixes
rdar://problem/20976169.