Driver: Specify the compilation phase to run to in crash reports
Rather than pretending to be invoked as `cpp`, we specify how much of
the compilation to run to generate intermediate files for a crash
report.

Internal only for now even though it's NFC, because it's only useful
for bitcode crash reports. The driver needs some refactoring to make
this useful elsewhere before its upstreamable.

rdar://problem/21436644