NFC: Change this code to match trunk.
Apple's master branch had diverged from trunk earlier, which then led to
merge conflicts, and then this part never got fixed up after the other
changes were pushed to trunk. The assignment to SuperType here is
redundant because the equivalent assignment is performed a few lines earlier,
and the assignment to IFace is equivalent to what is done on trunk.