Stop installing libclang from Apple Clang builds. rdar://problem/23099074
We're going to have swiftlang install it instead so that we can break the
rev-lock between swiftlang and clang. I'm doing this internally because
we're switching to CMake soon and it's not worth the effort to figure out
if we can change this on trunk. Most people don't use the BUILD_CLANG_ONLY
option, but someone could be relying on it.