Revert the part of r252690 that make typeof(bitfield) a hard error.
We have not good solution to write safe macros using typeof, and it seems like
using typeof to do this cannot really lead to any error. For now, allow this use
internally. We'll start a discussion upstream to see if this can be turned into
a warning that defaults to an error (but can then be disabled).