Emit clang.arc.use as part of the clean up for os_log builtin.
When we are emitting __builtin_os_log_format, check if any operand
is a retainable type, if it is and the emittd value is not a constant,
save the value in a vector.

At the end of emitting __builtin_os_log_format, push in a clang.arc.use
clean up for each value in the vector.

Since the corresponding objc_release clean ups are pushed in earlier,
the clang.arc.use will appear before the objc_release in the clean up
code.

This makes sure that objects will not be released until the end of
the full expression.

rdar://problem/24528966