Temporarily weaken __builtin_object_size. rdar://24701869
Improvements in clang's ability to track object sizes are causing problems
in code using our standard system headers when using _FORTIFY_SOURCE=2,
which is the default for our platforms. The failure mode is a runtime crash.
We can't fix the system headers without breaking compatibility. We're going
to investigate adding an attribute to identify structs that behave like
they have flexible array members, but there is no time to get that for
Everest. There is already an exception for arrays of size 0 or 1, so for
now, we can simply generalize that for any array size.