Actually fix darwin-ld.c failing test by disabling check for libclang_rt.ios.a
This check will fail if you don't have compiler-rt checked out and built before you run check-clang. This happens because adding libclang_rt.ios.a is gated on a check to ensure the library exists.

I'm not sure if that is the correct behavior or not, but it is long-standing and I'm not going to change it for this test case.