[indexstore] Use the stable enums for storage, but the C++ enums everywhere else
This commit standardizes the use of IndexSymbol (C++) vs. indexstore (C)
enums. We use the indexstore stable enums anywhere that either (1) is
part of the indexstore API, or (2) requires a stable value, such as in
the on-disk storage format. We use the IndexSymbol enums everywhere
else.

In practice, the changes are:
* We centralize all our conversion utilities to map between enum types
  into a new header IndexDataStoreSymbolUtils.h.
* IndexRecordWriter starts using the IndexSymbol enums in its interface
  and converts to indexstore enums internally for storage.
* IndexRecordReader actually maps to IndexSymbol values through an API
  instread of just casting them (this was already theoretically broken).
* IndexStore converts back to indexstore symbols through an API instead
  of just casting them (also already theoretically broken).

The upshot of this is that everything that has access to the C++ API
should be able to deal with the more convenient C++ enum class types,
and use the existing utility functions in IndexSymbol. However, it means
that if we add any non-clang symbol kinds, etc.to indexstore.h we also
have to duplicate it in IndexSymbol.h.