When compiling for C or Objective-C, push a normal lifetime.end cleanup instead of a NormalAndEH cleanup.
r265197 made changes to emit lifetime.end intrinsics in landing pads.
It also caused clang to emit two landing pads when the following code
including a call to objc_addExceptionHandler was compiled:

@try {
  uintptr_t token = objc_addExceptionHandler(&handler, nil);
  @throw [NSException new];
  objc_removeExceptionHandler(token);
} @catch (...) {

In this code, "handler" is expected to execute when the exception is
thrown, but that doesn't happen because the invoke instructions that
call objc_addExceptionHandler and @throw do not go to the same landing
pad.

rdar://problem/25763252