Strip clang with -S
It would be nice to use -Sx to make our binaries smaller, but without local symbols crash traces aren't symbolicated. This is creating extra burden on our bug screeners. We need to roll back this change for the clang-800 releases, and we can re-introduce it after we have sufficient tooling to symbolicate crashes attached to radars.

I've filed the following tracking radars:
<rdar://problem/26199865> BLOCKED: Strip clang with -Sx
<rdar://problem/26199769> Need tooling infastructure to symbolicate clang crashes in radars